chore: Move logic from `guards::UnauthenticatedServerManager` to `services::ServerManager`
