+++ service/src/services/mod.rs
@@ -8,5 +8,6 @@ pub mod invitation_service;
 pub mod jwt_service;
 pub(crate) mod live_xmpp_service;
 pub mod server_ctl;
+pub mod server_manager;
 pub mod user_service;
 pub mod xmpp_service;
